+++ app.js
@@ -8,35 +8,16 @@ const startTime = Date.now();
 require("dotenv").config();
 
 const express = require("express");
+const swaggerUi = require('swagger-ui-express');
 const cors = require("cors");
 const compression = require("compression");
-const helmet = require("helmet");
 const cookieParser = require("cookie-parser");
 const morgan = require("morgan");
+const helmet = require("helmet")
 const path = require("path");
-let expressOasGenerator;
-if (process.env.TEST_MODE) {
-    console.log("TEST MODE");
-    expressOasGenerator = require('express-oas-generator');
-}
-const app = express();
-if (process.env.TEST_MODE) {
-    const SPEC_OUTPUT_FILE_BEHAVIOR = {
-        PRESERVE: 'PRESERVE',
-        RECREATE: 'RECREATE'
-    };
-    expressOasGenerator.handleResponses(app, {
-        specOutputPath: './docs/generated.json',
-        alwaysServeDocs: true,
-        writeIntervalMs: 100,
-        specOutputFileBehavior: SPEC_OUTPUT_FILE_BEHAVIOR.RECREATE
-    });
-}
 const fs = require('fs');
-const http = require('http');
-const https = require('https');
-const swaggerUi = require('swagger-ui-express');
 
+const app = express();
 const cache = require('./config/cache');
 const nsfwModel = require("./models/NSFWModel");
 
@@ -137,8 +118,8 @@ app.get("/", (request, response) => {
 });
 
 const docsFolder = process.cwd() + "/docs/";
-let docs = require(docsFolder + 'generated.json');
-const os = require("os");
+let docs = require(docsFolder + 'docs.json');
+
 app.use("/docs", swaggerUi.serve, swaggerUi.setup(docs, {
     explorer: true,
 }));
@@ -160,88 +141,9 @@ app.get("*", function (req, res) {
 });
 
 
-//HTTP Server stuff here
-if (process.env.TEST_MODE) {
-    expressOasGenerator.handleRequests();
-}
-
-const httpPort = process.env.PORT || 5656;
-const httpsPort = process.env.PORT_HTTPS || 5657;
-try {
-    const httpServer = http.createServer(app);
-    httpServer.listen(httpPort, "0.0.0.0", back => {
-        console.log("Http server listening on port : " + httpPort)
-        console.log("http://localhost:" + httpPort)
-        //print all local ip
-        const interfaces = os.networkInterfaces();
-        for (const name of Object.keys(interfaces)) {
-            for (const net of interfaces[name]) {
-                console.log("http://" + net.address + ":" + httpPort)
-            }
-        }
-    });
-    let certsFolder = process.env.CERT_PATH || process.cwd() + '/certsFiles/';
-//end with /
-    if (!certsFolder.endsWith('/')) {
-        certsFolder = certsFolder + '/';
-    }
-    if (fs.existsSync(certsFolder)) {
-        try {
-
-            const credentials = {};
-            const certFilesName = ['certificate.crt', 'fullchain.pem'];
-            const keyFilesName = ['key.key', 'privkey.pem', 'private.key', 'privatekey.pem'];
-            const caFilesName = ['ca.crt', 'chain.pem', 'chain.cert.pem'];
-            for (const certFileName of certFilesName) {
-                if (fs.existsSync(certsFolder + certFileName)) {
-                    credentials.cert = fs.readFileSync(certsFolder + certFileName);
-                    console.log('cert file found : ' + certsFolder + certFileName);
-                    break;
-                }
-            }
-            if (!credentials.cert) {
-                console.error('cert file not found in : ' + certsFolder);
-            }
-            for (const keyFileName of keyFilesName) {
-                if (fs.existsSync(certsFolder + keyFileName)) {
-                    credentials.key = fs.readFileSync(certsFolder + keyFileName);
-                    console.log('key file found : ' + certsFolder + keyFileName);
-                    break;
-                }
-            }
-            if (!credentials.key) {
-                console.error('key file not found in : ' + certsFolder);
-            }
-            for (const caFileName of caFilesName) {
-                if (fs.existsSync(certsFolder + caFileName)) {
-                    credentials.ca = fs.readFileSync(certsFolder + caFileName);
-                    console.log('ca file found : ' + certsFolder + caFileName);
-                    break;
-                }
-            }
-            if (!credentials.ca) {
-                console.log('ca file not found in : ' + certsFolder);
-            }
-            const httpsServer = https.createServer(credentials, app);
-            httpsServer.listen(httpsPort, () => {
-                console.log("Https server listing on port : " + httpsPort)
-                console.log("https://localhost:" + httpsPort)
-            });
-        } catch (e) {
-            console.log("Can't start Https server");
-            console.log(e);
-        }
-    } else {
-        console.log("Can't start Https server, certs folder not found");
-    }
-} catch (e) {
-    console.log("Can't start Http server");
-    console.log(e);
-    if (!process.env.TEST_MODE) {
-        process.exit(1);
-    }
-}
 
 module.exports = {
-    app
+    app,
+    apiVersion,
+    testURL,
 }
