+++ docs/docs.json
@@ -14,244 +14,39 @@
     }
   ],
   "paths": {
-    "/api/v2/test": {
+    "/api/v3/health": {
       "get": {
-        "summary": "Test endpoint",
-        "parameters": {},
+        "summary": "Health Check",
+        "description": "Check if the API is up and running",
         "responses": {
           "200": {
-            "description": "Success",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "object",
-                  "properties": {
-                    "yes": {
-                      "type": "string",
-                      "description": "Test"
-                    }
-                  }
-                }
-              }
-            }
+            "description": "OK"
+          },
+          "500": {
+            "description": "Internal Server Error"
           }
         }
       }
     },
-    "/api/v2/categories": {
+    "/api/v3/meta/categories": {
       "get": {
-        "summary": "Get classification categories",
-        "parameters": [],
+        "summary": "Get Categories",
+        "description": "Get all categories",
         "responses": {
           "200": {
-            "description": "Success",
+            "description": "OK",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "array",
                   "items": {
-                    "type": "array",
-                    "items": {
-                      "type": "string",
-                      "description": "Category"
+                    "type": "object",
+                    "properties": {
+                      "class": {
+                        "type": "string"
+                      }
                     }
                   }
-                }
-              }
-            }
-          }
-        }
-      }
-    },
-    "/api/v2/hosts": {
-      "get": {
-        "summary": "Get hosts filter list",
-        "parameters": [],
-        "responses": {
-          "200": {
-            "description": "Success",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "object",
-                  "properties": {
-                    "allowedHost": {
-                      "type": "array",
-                      "items": {
-                        "type": "string",
-                        "description": "Host"
-                      }
-                    },
-                    "blockedHost": {
-                      "type": "array",
-                      "items": {
-                        "type": "string",
-                        "description": "Host"
-                      }
-                    },
-                    "allowedAll": {
-                      "type": "boolean",
-                      "description": "Allow all hosts"
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    },
-    "/api/v2/classification/hash": {
-      "post": {
-        "summary": "Get classification for hash",
-        "parameters": {
-          "body": {
-            "description": "Base64 encoded image hash sha256",
-            "required": true,
-            "schema": {
-              "type": "string",
-              "format": "base64"
-            }
-          }
-        },
-        "responses": {
-          "200": {
-            "description": "Cache found",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/classification"
-                }
-              }
-            }
-          },
-          "404": {
-            "description": "Cache not found",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/cache/miss"
-                }
-              }
-            }
-          }
-        }
-      }
-    },
-    "/api/v2/classification/hash/{hash}": {
-      "get": {
-        "summary": "Get classification for hash",
-        "parameters": {
-          "hash": {
-            "description": "Image hash sha256",
-            "required": true,
-            "schema": {
-              "type": "string",
-              "format": "base64"
-            }
-          }
-        },
-        "responses": {
-          "200": {
-            "description": "Cache found",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/classification"
-                }
-              }
-            }
-          },
-          "404": {
-            "description": "Cache not found",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/cache/miss"
-                }
-              }
-            }
-          }
-        }
-      }
-    },
-    "/api/v2/classification": {
-      "post": {
-        "summary": "Get classification for image",
-        "parameters": {
-          "body": {
-            "description": "Image",
-            "required": true,
-            "schema": {
-              "type": "string",
-              "format": "base64"
-            }
-          }
-        },
-        "responses": {
-          "201": {
-            "description": "Classification found",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/classification"
-                }
-              }
-            }
-          },
-          "406": {
-            "description": "Invalid request",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/classification/error"
-                }
-              }
-            }
-          },
-          "500": {
-            "description": "Internal error",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/classification/error"
-                }
-              }
-            }
-          }
-        }
-      }
-    },
-    "/api/v2/classification/{url}": {
-      "get": {
-        "summary": "Get classification for url",
-        "parameters": {
-          "url": {
-            "description": "Image url",
-            "required": true,
-            "schema": {
-              "type": "string",
-              "format": "url"
-            }
-          }
-        },
-        "responses": {
-          "200": {
-            "description": "Classification found",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/classification"
-                }
-              }
-            }
-          },
-          "406": {
-            "description": "Invalid request",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/classification/error"
                 }
               }
             }
@@ -268,17 +63,6 @@
           "model": {
             "type": "string"
           }
-        },
-        "error": {
-          "type": "object",
-          "properties": {
-            "error": {
-              "type": "string"
-            },
-            "status": {
-              "type": "integer"
-            }
-          }
         }
       },
       "cache": {
@@ -286,14 +70,6 @@
         "properties": {
           "model": {
             "type": "string"
-          }
-        },
-        "miss": {
-          "type": "object",
-          "properties": {
-            "error": {
-              "type": "string"
-            }
           }
         }
       }
