+++ package.json
@@ -4,7 +4,7 @@
   "description": "Node.js + Express API for classifying images",
   "main": "app.js",
   "scripts": {
-    "start": "node app.js",
+    "start": "node server.js",
     "test": "node test.js"
   },
   "dependencies": {
@@ -18,7 +18,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "express-oas-generator": "^1.0.45",
+
     "helmet": "^6.0.0",
     "ioredis": "^5.2.3",
     "is-ci": "^3.0.1",
@@ -37,5 +37,10 @@
     "node",
     "glitch",
     "express"
-  ]
+  ],
+  "devDependencies": {
+    "jest": "^29.2.2",
+    "superagent": "^8.0.3",
+    "supertest": "^6.3.1"
+  }
 }
